Move creation functions to cartridgeview
